pyproject.toml
[tool.poetry]
name = "starnavi-test-task"
version = "0.1.0"
description = "API for workflow management"
authors = ["Dmytro Petliovanyi <petliovany@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pre-commit = "^3.7.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
isort = "^5.13.2"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
alembic = "^1.13.1"
pydantic = "^2.6.4"
networkx = "^3.3"
matplotlib = "^3.8.4"
asyncpg = "^0.29.0"
types-redis = "^4.6.0.20240409"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
psycopg2 = "^2.9.9"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
asyncpgsa = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
